Move getAuthors, parseNames to DryadWorkflowUtils
