+++ dspace/modules/api/src/main/java/org/dspace/workflow/DryadWorkflowUtils.java
@@ -20,6 +20,9 @@ import java.io.IOException;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.StringTokenizer;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 /**
  * User: kevin (kevin at atmire.com)
@@ -156,6 +159,83 @@ public class DryadWorkflowUtils {
 
         }
         return false;
+    }
+
+    public static String getAuthors(Item item) {
+        ArrayList<DCValue> mdlist = new ArrayList<DCValue>();
+        for (DCValue i : item.getMetadata("dc.contributor.author")) {
+            mdlist.add(i);
+        }
+        for (DCValue i : item.getMetadata("dc.creator")) {
+            mdlist.add(i);
+        }
+        for (DCValue i : item.getMetadata("dc.contributor")) {
+            mdlist.add(i);
+        }
+        StringBuilder buffer = new StringBuilder();
+        buffer.append(parseName((DCValue[]) mdlist.toArray(new DCValue[1])));
+
+        String author = buffer.toString().trim() + " ";
+        return author;
+    }
+
+    private static String parseName(DCValue[] aMetadata) {
+        StringBuilder buffer = new StringBuilder();
+        int position = 0;
+
+        for (DCValue metadata : aMetadata) {
+            StringBuilder authorString = new StringBuilder();
+            authorString.append("@");
+            if (metadata.value.indexOf(",") != -1) {
+                String[] parts = metadata.value.split(",");
+
+                if (parts.length > 1) {
+                    StringTokenizer tokenizer = new StringTokenizer(parts[1], ". ");
+
+                    authorString.append(parts[0]).append(" ");
+
+                    while (tokenizer.hasMoreTokens()) {
+                        authorString.append(tokenizer.nextToken().charAt(0));
+                    }
+                }
+            } else {
+                // now the minority case (as we clean up the data)
+                String[] parts = metadata.value.split("\\s+|\\.");
+                String author = parts[parts.length - 1].replace("\\s+|\\.", "");
+                char ch;
+
+                authorString.append(author).append(" ");
+
+                for (int index = 0; index < parts.length - 1; index++) {
+                    if (parts[index].length() > 0) {
+                        ch = parts[index].replace("\\s+|\\.", "").charAt(0);
+                        authorString.append(ch);
+                    }
+                }
+            }
+            authorString.append("@");
+
+            // check for orcid:
+            if (metadata.authority != null) {
+                Pattern p = Pattern.compile(".+orcid::(.+)");
+                Matcher m = p.matcher(metadata.authority);
+                if (m.matches()) {
+                    authorString.append("#").append(m.group(1)).append("#");
+                }
+            }
+
+            buffer.append(authorString.toString());
+
+            if (++position < aMetadata.length) {
+                if (aMetadata.length > 2) {
+                    buffer.append(", ");
+                } else {
+                    buffer.append(" and ");
+                }
+            }
+        }
+
+        return buffer.length() > 0 ? buffer.toString() : "";
     }
